Change `pktId` from Int to Long in database and repository
This commit updates the `pktId` field in the `Message` entity and related database operations to use `Long` instead of `Int`. This change accommodates larger packet identifiers.

Additionally, the following changes were made:
- Added a check in `BleMeshRepository.persist` to prevent duplicate message insertion based on `pktId`.
- Added logging for `pktId` in `BleMeshRepository` send and persist functions.
- Disabled backup and full backup content in `AndroidManifest.xml`.
- Incremented the database version to 6.
- Added `identity-jvm` dependency.

+++ core-database/src/main/java/com/example/disastermesh/core/database/entities/Message.kt
@@ -5,6 +5,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
+import com.android.identity.cbor.Uint
 
 /**
  * One message inside a chat.
@@ -22,7 +23,7 @@ import androidx.room.PrimaryKey
 )
 data class Message(
     @PrimaryKey(autoGenerate = true) val msgId: Long = 0,
-    val pktId : Int? = null,
+    val pktId : Long? = null,
     val chatId : Long,
     val mine   : Boolean,                  // true = I sent it
     val ts     : Long = System.currentTimeMillis(),
